fix: stop asserting Brazos cap type from a flag that can't say what it is
evaluate_cap_status() decoded AssessmentHistory.cap_account as if it
were always HCAD's Y/N/Pending homestead-cap flag. Brazos's cap_account
is a derived signal (Y when appraised_value > assessed_value for that
year) that only says a capping reduction happened, not which cap
applied -- so Brazos rows were silently asserted to be homestead-capped
(10%) or circuit-breaker-capped (20%) with no basis in BCAD's data,
feeding a fabricated limit straight into ARB evidence. This was already
live against the 705,560 Brazos AssessmentHistory rows loaded this
session (issue #13).

Moves evaluate_cap_status out of counties/harris/assessment_history.py
into counties/common/cap_status.py -- it's genuinely used by both
counties, not something counties/common/history.py should have to
lazy-import out of a single county's package. Adds
COUNTIES_WITH_TYPED_CAP_FLAG so only counties whose flag actually
distinguishes cap type get a homestead/circuit_breaker answer; every
other county gets an honest cap_type="unknown" (still with a real
increase_percent) instead of a guess.

Surfaces from the architecture review's top recommendation
(/improve-codebase-architecture, 2026-08-09). Narrows, but does not
close, issue #14 -- the tax-year/SB2 effective-date and value-ceiling
gaps in the homestead/circuit-breaker rule itself are still open.
