+++ counties/common/history.py
@@ -10,15 +10,17 @@ from __future__ import annotations
 from typing import Any
 
 from counties.common.analysis import year_over_year_percent
+from counties.common.cap_status import evaluate_cap_status
 
 
 def assessment_history_rows(
     account_number: str, county: str = "harris", limit: int = 5
 ) -> list[dict[str, Any]]:
     """Per-year assessed values, newest first, with YoY change and cap status."""
-    # Imported lazily: this module is part of the county-neutral layer, and
-    # importing a county app at module scope would invert that dependency.
-    from counties.harris.assessment_history import evaluate_cap_status
+    # Imported lazily: AssessmentHistory is a Harris-app model shared via its
+    # county column (wayfinder ticket #9); importing a county app at module
+    # scope would invert the county-neutral layer's dependency direction.
+    # evaluate_cap_status is genuinely part of this layer -- see cap_status.py.
     from counties.harris.models import AssessmentHistory
 
     history = list(
